+++ src/components/SettingsMapsforgeControl.tsx
@@ -1,0 +1,104 @@
+/**
+ * External dependencies
+ */
+import React, {
+	useContext,
+    useEffect,
+    useRef,
+    useState,
+} from 'react';
+import {
+	View,
+} from 'react-native';
+import {
+	Icon,
+    Text,
+} from 'react-native-paper';
+import { useTranslation } from 'react-i18next';
+import { get } from 'lodash-es';
+
+/**
+ * Internal dependencies
+ */
+import { AppContext } from '../Context';
+import { MapSettings, MapsforgeGeneral } from '../types';
+import ListItemModalControl from './ListItemModalControl';
+import { defaults } from '../constants';
+import { NumericRowControl } from './NumericRowControls';
+import IconIcomoon from './IconIcomoon';
+
+const SettingsMapsforgeControl = () => {
+
+	const { t } = useTranslation();
+
+	const {
+		mapSettings,
+		setMapSettings,
+	} = useContext( AppContext );
+
+	const [value,setValue] = useState<MapsforgeGeneral>( get( mapSettings, ['mapsforgeGeneral'], defaults.mapSettings.mapsforgeGeneral ) );
+	const valueRef = useRef( value );
+    useEffect( () => {
+        valueRef.current = value;
+    }, [value] );
+    const save = () => {
+
+        console.log( 'debug valueRef.current', valueRef.current ); // debug
+
+        return setMapSettings && setMapSettings( ( mapSettings: MapSettings ) => ( {
+            ...mapSettings,
+            ...( valueRef.current && { mapsforgeGeneral: valueRef.current } ),
+        } ) );
+    }
+    useEffect( () => save, [] );    // Save on unmount.
+
+	return <ListItemModalControl
+		anchorLabel={ t( 'settings.mapsforgeGeneral' ) }
+        anchorIcon={ props => <IconIcomoon size={ 25 } name="mapsforge_puzzle_cog" {...props} /> }
+		header={ t( 'settings.mapsforgeGeneral' ) }
+		hasHeaderBackPress={ true }
+	>
+
+        <Text style={ { marginBottom: 10 } }>{ t( 'hint.applyToAllMapsforge' ) }</Text>
+        <Text style={ { marginBottom: 10 } }>{ t( 'hint.changeNeedsRestart' ) }</Text>
+
+        <NumericRowControl
+            label={ t( 'lineScale' ) }
+            optKey={ 'lineScale' }
+            numType={ 'float' }
+            options={ value }
+            setOptions={ ( newValue ) => {
+                setValue( newValue );
+            } }
+            validate={ val => val >= 0 }
+            Info={ t( 'hint.maps.lineScale' ) }
+        />
+
+        <NumericRowControl
+            label={ t( 'textScale' ) }
+            optKey={ 'textScale' }
+            numType={ 'float' }
+            options={ value }
+            setOptions={ ( newValue ) => {
+                setValue( newValue );
+            } }
+            validate={ val => val >= 0 }
+            Info={ t( 'hint.maps.textScale' ) }
+        />
+
+        <NumericRowControl
+            label={ t( 'symbolScale' ) }
+            optKey={ 'symbolScale' }
+            numType={ 'float' }
+            options={ value }
+            setOptions={ ( newValue ) => {
+                setValue( newValue );
+            } }
+            validate={ val => val >= 0 }
+            Info={ t( 'hint.maps.symbolScale' ) }
+        />
+
+	</ListItemModalControl>;
+};
+
+export default SettingsMapsforgeControl;